+++ tests/testthat/test_feature_representation.R
@@ -2,14 +2,14 @@ context("feature_representation")
 
 test_that("numeric", {
   # simulate data
-  pu <- data.frame(id = seq_len(10), cost = c(0.2, NA, runif(8)),
+  pu <- data.frame(id = seq_len(10), cost = c(0.2, NA_real_, runif(8)),
                    spp1 = runif(10), spp2 = c(rpois(9, 4), NA))
   # create problem
   x <- problem(pu$cost, data.frame(id = seq_len(2), name = c("spp1", "spp2")),
                as.matrix(t(pu[, 3:4])))
   # create a solution
   y <- rep(c(0, 1), 5)
-  y[is.na(pu$cost)] <- 0
+  y[is.na(pu$cost)] <- NA_real_
   attr(y, "runtime") <- 5
   # calculate representation
   r <- feature_representation(x, y)
@@ -24,11 +24,12 @@ test_that("numeric", {
                                          sum(pu$spp2, na.rm  = TRUE)))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("matrix (single zone)", {
   # simulate data
-  pu <- data.frame(id = seq_len(10), cost = c(0.2, runif(9)),
+  pu <- data.frame(id = seq_len(10), cost = c(0.2, NA_real_, runif(8)),
                    spp1 = runif(10), spp2 = c(rpois(9, 4), NA))
   # create problem
   x <- problem(matrix(pu$cost, ncol = 1),
@@ -36,7 +37,7 @@ test_that("matrix (single zone)", {
                as.matrix(t(pu[, 3:4])))
   # create a solution
   y <- matrix(rep(c(0, 1), 5), ncol = 1)
-  y[is.na(pu$cost)] <- 0
+  y[is.na(pu$cost)] <- NA_real_
   # calculate representation
   r <- feature_representation(x, y)
   # create correct result
@@ -50,6 +51,7 @@ test_that("matrix (single zone)", {
                                          sum(pu$spp2, na.rm  = TRUE)))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("matrix (multiple zones)", {
@@ -64,7 +66,7 @@ test_that("matrix (multiple zones)", {
                list(as.matrix(t(pu[, 4:5])), as.matrix(t(pu[, 6:7]))))
   # create a solution
   y <- matrix(c(rep(c(0, 0.5), 5), rep(c(0.5, 0), 5)), ncol = 2)
-  y[is.na(as.matrix(pu[, c("cost_1", "cost_2")]))] <- 0
+  y[is.na(as.matrix(pu[, c("cost_1", "cost_2")]))] <- NA_real_
   # calculate representation
   r <- feature_representation(x, y)
   # create correct result
@@ -82,6 +84,7 @@ test_that("matrix (multiple zones)", {
                                       sum(pu$spp2_2, na.rm  = TRUE)))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("data.frame (single zone)", {
@@ -92,7 +95,7 @@ test_that("data.frame (single zone)", {
   x <- problem(pu, c("spp1", "spp2"), cost_column = "cost")
   # create a solution
   y <- data.frame(solution = rep(c(0, 1), 5))
-  y[[1]][is.na(pu$cost)] <- 0
+  y[[1]][is.na(pu$cost)] <- NA_real_
   # calculate representation
   r <- feature_representation(x, y)
   # create correct result
@@ -104,6 +107,7 @@ test_that("data.frame (single zone)", {
                                       sum(pu$spp2, na.rm  = TRUE)))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("data.frame (multiple zone)", {
@@ -117,8 +121,8 @@ test_that("data.frame (multiple zone)", {
                c("cost_1", "cost_2"))
   # create a solution
   y <- data.frame("z1" = rep(c(0, 0.5), 5), "z2" = rep(c(0.5, 0), 5))
-  y[[1]][is.na(pu$cost_1)] <- 0
-  y[[2]][is.na(pu$cost_2)] <- 0
+  y[[1]][is.na(pu$cost_1)] <- NA_real_
+  y[[2]][is.na(pu$cost_2)] <- NA_real_
   # calculate representation
   r <- feature_representation(x, y)
   # create correct result
@@ -136,6 +140,7 @@ test_that("data.frame (multiple zone)", {
                                       sum(pu$spp2_2, na.rm  = TRUE)))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("Spatial (single zone)", {
@@ -144,7 +149,7 @@ test_that("Spatial (single zone)", {
   pu <- sim_pu_polygons
   pu$cost[1:5] <- NA
   pu$solution <- rep(c(0, 1), 5)
-  pu$solution[is.na(pu$cost)] <- 0
+  pu$solution[is.na(pu$cost)] <- NA_real_
   pu$spp1 <- runif(10)
   pu$spp2 <- c(rpois(9, 1), NA)
   # create problem
@@ -164,6 +169,7 @@ test_that("Spatial (single zone)", {
                                       sum(pu$spp2, na.rm  = TRUE)))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("Spatial (multiple zone)", {
@@ -177,8 +183,8 @@ test_that("Spatial (multiple zone)", {
   pu$spp2_2 <- rpois(nrow(pu), 1)
   pu$s1 <- rep(c(0, 0.5), nrow(pu) / 2)
   pu$s2 <- rep(c(0.5, 0), nrow(pu) / 2)
-  pu$s1[is.na(pu$cost_1)] <- 0
-  pu$s2[is.na(pu$cost_2)] <- 0
+  pu$s1[is.na(pu$cost_1)] <- NA_real_
+  pu$s2[is.na(pu$cost_2)] <- NA_real_
   # create problem
   x <- problem(pu,
                zones(z1 = c("spp1_1", "spp2_1"), z2 = c("spp1_2", "spp2_2"),
@@ -203,6 +209,7 @@ test_that("Spatial (multiple zone)", {
                                       sum(pu$spp2_2, na.rm  = TRUE)))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("Raster (single zone)", {
@@ -213,7 +220,7 @@ test_that("Raster (single zone)", {
   # create a solution
   y <- raster::setValues(sim_pu_raster,
                          rep(c(0, 1), raster::ncell(sim_pu_raster) / 2))
-  y[is.na(sim_pu_raster)] <- 0
+  y[is.na(sim_pu_raster)] <- NA_real_
   # calculate representation
   r <- feature_representation(x, y)
   # create correct result
@@ -226,6 +233,7 @@ test_that("Raster (single zone)", {
     relative_held = absolute_held / raster::cellStats(sim_features, "sum"))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("Raster (multiple zone)", {
@@ -241,9 +249,9 @@ test_that("Raster (multiple zone)", {
                       rep(c(0.3, 0), raster::ncell(sim_pu_raster) / 2)),
     raster::setValues(sim_pu_raster,
                       rep(c(0.4, 0), raster::ncell(sim_pu_raster) / 2)))
-  y[[1]][is.na(sim_pu_zones_stack[[1]])] <- 0
-  y[[2]][is.na(sim_pu_zones_stack[[2]])] <- 0
-  y[[3]][is.na(sim_pu_zones_stack[[3]])] <- 0
+  y[[1]][is.na(sim_pu_zones_stack[[1]])] <- NA_real_
+  y[[2]][is.na(sim_pu_zones_stack[[2]])] <- NA_real_
+  y[[3]][is.na(sim_pu_zones_stack[[3]])] <- NA_real_
   # calculate representation
   r <- feature_representation(x, y)
   # create correct result
@@ -283,6 +291,7 @@ test_that("Raster (multiple zone)", {
                       raster::cellStats(sim_features_zones[[3]], "sum")))
   # run tests
   expect_equal(r, r2)
+  expect_equal(nrow(na.omit(r)), nrow(r))
 })
 
 test_that("invalid inputs", {